Create dedicated format wrapper
Move on from the polyglot-because-its-fun wrapper binary. We want to
achieve something with purpose so let us do exactly that.
