+++ bin/polywrap/src/stage-snippet-load-URI-b64.js
